Third commit

+++ src/main/java/br/com/rj/marvel/repository/CharacterRepository.java
@@ -7,6 +7,6 @@ import br.com.rj.marvel.model.Character;
 import br.com.rj.marvel.repository.character.CharacterRepositoryQuery;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Long>, CharacterRepositoryQuery {
+public interface CharacterRepository extends JpaRepository<Character, Integer>, CharacterRepositoryQuery {
 
 }
